% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\docType{data}
\name{xCell.data}
\alias{xCell.data}
\title{xCell datasets}
\format{list:
\describe{
  \item{spill}{spillover matrix and calibration parameters}
  \item{signatures}{the signatures for calculating scores}
  \item{genes}{genes to use to calculate xCell}
}}
\usage{
xCell.data
}
\description{
xCell datasets
}
\keyword{datasets}
