% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\docType{package}
\name{xCell}
\alias{xCell}
\alias{xCell-package}
\alias{xCell}
\title{xCell: A package for calculating cell type enrichments.}
\format{list:
\describe{
  \item{spill}{spillover matrix and calibration parameters}
  \item{signatures}{the signatures for calculating scores}
  \item{genes}{genes to use to calculate xCell}
}}
\usage{
xCell
}
\description{
xCell: A package for calculating cell type enrichments.

xCell datasets
}
\keyword{datasets}
