% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{xCellAnalysis}
\alias{xCellAnalysis}
\title{The xCell analysis pipeline}
\usage{
xCellAnalysis(expr, signatures = NULL, genes = NULL, spill = NULL,
  file.name = NULL, rnaseq = TRUE, calibration = NULL, alpha = 1,
  save.raw = FALSE)
}
\arguments{
\item{expr}{the gene expression data set. A matrix with row names as symbols and columns as samples.}

\item{signatures}{a GMT object of signatures.}

\item{genes}{list of genes to use in the analysis.}

\item{spill}{the Spillover object for adjusting the scores.}

\item{file.name}{string for the file name for saving the scores. Default is NULL.}

\item{rnaseq}{TRUE if the expression data is RNA-seq. Default = TRUE}

\item{calibration}{a calibration value to override the spillover calibration parameters. Default = NULL}

\item{alpha}{a value to override the spillover alpha parameter. Deafult = 1}

\item{save.raw}{TRUE to save a raw}
}
\value{
the adjusted xCell scores
}
\description{
\code{xCellAnalysis} Returns the xCell cell types enrichment scores.
}
