% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{xCellSignifcanceBetaDist}
\alias{xCellSignifcanceBetaDist}
\title{Calculate significance p-values for the null hypothesis that the cell type is not present in the mixture using a random matrix.}
\usage{
xCellSignifcanceBetaDist(scores, beta_params = NULL, rnaseq = T,
  file.name = NULL)
}
\arguments{
\item{scores}{the xCell scores.}

\item{file.name}{file name to write the p-values table.}

\item{beta_dist}{the pre-calculated beta distribution from random mixtures.}
}
\value{
a p-values matrix for each score.
}
\description{
\code{xCellSignifcanceBetaDist} Returns the FDR adjusted p-values of the chance that the cell is not present in the mixture.
}
