% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell.R
\name{xCellAnalysis}
\alias{xCellAnalysis}
\title{The xCell analysis pipeline}
\usage{
xCellAnalysis(expr, signatures = NULL, genes = NULL, spill = NULL,
  rnaseq = TRUE, file.name = NULL, scale = TRUE, alpha = 0.5,
  save.raw = FALSE, parallel.sz = 4)
}
\arguments{
\item{expr}{the gene expression data set. A matrix with row names as symbols and columns as samples.}

\item{signatures}{a GMT object of signatures.}

\item{genes}{list of genes to use in the analysis.}

\item{spill}{the Spillover object for adjusting the scores.}

\item{rnaseq}{if true than use RNAseq spillover and calibration paramters, else use array parameters.}

\item{file.name}{string for the file name for saving the scores. Default is NULL.}

\item{scale}{if TRUE, uses scaling to trnasform scores using fit.vals}

\item{alpha}{a value to override the spillover alpha parameter. Deafult = 0.5}

\item{save.raw}{TRUE to save a raw}

\item{parallel.sz}{integer for the number of threads to use. Default is 4.}
}
\value{
the adjusted xCell scores
}
\description{
\code{xCellAnalysis} Returns the xCell cell types enrichment scores.
}
